# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Formly - Online Form Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Form {
  id: string;
  title: string;
  description?: string | null;
  themeColor: string;
  isPublished: boolean;
  questionCount: number;
  responseCount: number;
  createdAt: string;
  updatedAt: string;
}

export type QuestionType = (typeof QuestionType)[keyof typeof QuestionType];

export const QuestionType = {
  short_text: "short_text",
  long_text: "long_text",
  multiple_choice: "multiple_choice",
  checkbox: "checkbox",
  dropdown: "dropdown",
  ranking: "ranking",
  rating: "rating",
  opinion_scale: "opinion_scale",
  date: "date",
  email: "email",
  phone: "phone",
  number: "number",
  yes_no: "yes_no",
} as const;

export type LogicRuleCondition =
  (typeof LogicRuleCondition)[keyof typeof LogicRuleCondition];

export const LogicRuleCondition = {
  equals: "equals",
  not_equals: "not_equals",
  contains: "contains",
  is_empty: "is_empty",
  is_not_empty: "is_not_empty",
  greater_than: "greater_than",
  less_than: "less_than",
} as const;

export interface LogicRule {
  condition: LogicRuleCondition;
  value?: string | null;
  jumpToQuestionId?: string | null;
  jumpToEnd?: boolean;
}

export interface Question {
  id: string;
  formId: string;
  type: QuestionType;
  title: string;
  description?: string | null;
  required: boolean;
  order: number;
  options?: string[] | null;
  logic?: LogicRule[] | null;
  createdAt: string;
}

export type FormWithQuestions = Form & {
  questions: Question[];
};

export interface CreateFormBody {
  title: string;
  description?: string;
  themeColor?: string;
}

export interface UpdateFormBody {
  title?: string;
  description?: string;
  themeColor?: string;
  isPublished?: boolean;
}

export type CreateQuestionBodyType =
  (typeof CreateQuestionBodyType)[keyof typeof CreateQuestionBodyType];

export const CreateQuestionBodyType = {
  short_text: "short_text",
  long_text: "long_text",
  multiple_choice: "multiple_choice",
  checkbox: "checkbox",
  dropdown: "dropdown",
  ranking: "ranking",
  rating: "rating",
  opinion_scale: "opinion_scale",
  date: "date",
  email: "email",
  phone: "phone",
  number: "number",
  yes_no: "yes_no",
} as const;

export interface CreateQuestionBody {
  type: CreateQuestionBodyType;
  title: string;
  description?: string;
  required?: boolean;
  options?: string[];
  logic?: LogicRule[];
}

export type UpdateQuestionBodyType =
  (typeof UpdateQuestionBodyType)[keyof typeof UpdateQuestionBodyType];

export const UpdateQuestionBodyType = {
  short_text: "short_text",
  long_text: "long_text",
  multiple_choice: "multiple_choice",
  checkbox: "checkbox",
  dropdown: "dropdown",
  ranking: "ranking",
  rating: "rating",
  opinion_scale: "opinion_scale",
  date: "date",
  email: "email",
  phone: "phone",
  number: "number",
  yes_no: "yes_no",
} as const;

export interface UpdateQuestionBody {
  type?: UpdateQuestionBodyType;
  title?: string;
  description?: string;
  required?: boolean;
  options?: string[];
  logic?: LogicRule[];
}

export interface Response {
  id: string;
  formId: string;
  completed: boolean;
  submittedAt: string;
}

export interface Answer {
  id: string;
  questionId: string;
  questionTitle: string;
  value?: string | null;
}

export type ResponseWithAnswers = Response & {
  answers: Answer[];
};

export type SubmitResponseBodyAnswersItem = {
  questionId: string;
  value?: string;
};

export interface SubmitResponseBody {
  answers: SubmitResponseBodyAnswersItem[];
  completed?: boolean;
}

export interface PaginatedResponses {
  data: ResponseWithAnswers[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export type FormStatsResponsesPerDayItem = {
  date: string;
  count: number;
};

export type QuestionStatTopAnswersItem = {
  value: string;
  count: number;
};

export interface QuestionStat {
  questionId: string;
  questionTitle: string;
  type: string;
  answerCount: number;
  topAnswers: QuestionStatTopAnswersItem[];
}

export interface FormStats {
  formId: string;
  totalResponses: number;
  completedResponses: number;
  completionRate: number;
  responsesPerDay: FormStatsResponsesPerDayItem[];
  questionStats: QuestionStat[];
}

export interface DashboardSummary {
  totalForms: number;
  publishedForms: number;
  totalResponses: number;
  responsesThisWeek: number;
  recentForms: Form[];
}

export interface SheetIntegration {
  id: string;
  formId: string;
  spreadsheetId: string;
  spreadsheetName: string;
  sheetName: string;
  enabled: boolean;
  lastSyncedAt?: string | null;
  createdAt: string;
}

export interface SaveSheetIntegrationBody {
  spreadsheetId: string;
  spreadsheetName: string;
  sheetName: string;
  enabled?: boolean;
}

export type PublishFormBody = {
  published: boolean;
};

export type ReorderQuestionsBody = {
  questionIds: string[];
};

export type ListResponsesParams = {
  /**
   * @minimum 1
   */
  page?: number;
  /**
   * @minimum 1
   * @maximum 100
   */
  limit?: number;
};

export type SyncToSheets200 = {
  rowsSynced: number;
  message: string;
};