# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Formly - Online Form Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /forms/{id}/stats:
    get:
      operationId: getFormStats
      tags: [dashboard]
      summary: Get analytics stats for a form
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Form statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormStats"
...
components:
  schemas:
...
    FormStats:
      type: object
      properties:
        formId:
          type: string
        totalResponses:
          type: integer
        completedResponses:
          type: integer
        completionRate:
          type: number
        responsesPerDay:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              count:
                type: integer
            required: [date, count]
        questionStats:
          type: array
          items:
            $ref: "#/components/schemas/QuestionStat"
      required: [formId, totalResponses, completedResponses, completionRate, responsesPerDay, questionStats]
    QuestionStat:
      type: object
      properties:
        questionId:
          type: string
        questionTitle:
          type: string
        type:
          type: string
        answerCount:
          type: integer
        topAnswers:
          type: array
          items:
            type: object
            properties:
              value:
                type: string
              count:
                type: integer
            required: [value, count]
      required: [questionId, questionTitle, type, answerCount, topAnswers]
...